test: split vsock host behavior tests
